backend/requirements.txt
# FastAPI and server
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
pydantic>=2.6.1
pydantic-settings>=2.1.0

# LangChain and LLM
langchain>=0.1.6
langchain-google-genai>=0.0.6
langchain-community>=0.0.19
langchain-core>=0.1.23
langchain-text-splitters

# Vector store
chromadb>=0.4.22

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.1
aiofiles>=23.2.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
httpx>=0.26.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
